Ubuntu 12 contribution / "porting mode" programming - many IDEs in different platforms and OSes... hellomultiverse contribution...
Eu te amo demais Thommy Gustavo Borregaard Sørensen!!
Eu te amo muito Thomas!!

Mummi contribution!!! Feliz 61 anos!!

porting to Linux Ubuntu using Linux Ubuntu 12.04 LTS as VMware guest at inha at running Windows 7 at veriverse Workgroup at Home at Alameda Princesa Izabel 3000 Curitiba Paraná Brazil!!



｛
｝

                  -   -   -
Mummi contribution!!, eu te amo demais!!

-----8nnnnnnnn8
++++++++++

+++ os/lnx_application.cpp
@@ -694,7 +694,7 @@ if(__get_module_state()->m_pmapHWND == NULL)
    #endif
    }
 
-   unsigned long application::get_thread_id()
+   uint32_t application::get_thread_id()
    {
       return ::pthread_self();
    }
